move all logging into base estimator logdir
