+++ gluoncv/auto/estimators/center_net.py
@@ -10,14 +10,13 @@ from mxnet import autograd
 from ...data.transforms.presets.center_net import CenterNetDefaultTrainTransform
 from ...data.transforms.presets.center_net import CenterNetDefaultValTransform, get_post_transform
 from ...data.batchify import Tuple, Stack, Pad
-from ...utils.metrics.accuracy import Accuracy
 from ...utils import LRScheduler, LRSequential
 from ...model_zoo.center_net import get_center_net, get_base_network
 from ...loss import MaskedL1Loss, HeatmapFocalLoss
 
 from ..data.coco_detection import coco_detection, load_coco_detection
-from .base_estimator import BaseEstimator, set_default, DotDict
-from .common import train_hp, valid_hp
+from .base_estimator import BaseEstimator, set_default, ConfigDict
+from .common import train, validation
 
 from sacred import Experiment, Ingredient
 
@@ -41,7 +40,7 @@ def center_net_default():
     center_reg_weight = 1.0  # Center regression loss weight
     data_shape = (512, 512)
 
-@train_hp.config
+@train.config
 def update_train_config():
     gpus = (0, 1, 2, 3, 4, 5, 6, 7)
     pretrained_base = True  # whether load the imagenet pre-trained base
@@ -55,7 +54,7 @@ def update_train_config():
     warmup_epochs = 0  # number of warmup epochs
 
 
-@valid_hp.config
+@validation.config
 def update_valid_config():
     flip_test = True  # use flip in validation test
     nms_thresh = 0  # 0 means disable
@@ -63,7 +62,7 @@ def update_valid_config():
     post_nms = 100  # post nms topk
 
 ex = Experiment('center_net_default',
-                ingredients=[coco_detection, train_hp, valid_hp, center_net])
+                ingredients=[coco_detection, train, validation, center_net])
 
 @ex.config
 def default_configs():
@@ -90,7 +89,7 @@ class CenterNetEstimator(BaseEstimator):
             raise NotImplementedError
 
         # network
-        ctx = [mx.gpu(int(i)) for i in self._cfg.train_hp.gpus]
+        ctx = [mx.gpu(int(i)) for i in self._cfg.train.gpus]
         ctx = ctx if ctx else [mx.cpu()]
         self._ctx = ctx
         net_name = '_'.join(('center_net', self._cfg.center_net.base_network, self._cfg.dataset))
@@ -98,7 +97,7 @@ class CenterNetEstimator(BaseEstimator):
             ('heatmap', {'num_output': train_dataset.num_class, 'bias': self._cfg.center_net.heads.bias}),
             ('wh', {'num_output': self._cfg.center_net.heads.wh_outputs}),
             ('reg', {'num_output': self._cfg.center_net.heads.reg_outputs})])
-        base_network = get_base_network(self._cfg.center_net.base_network, pretrained=self._cfg.train_hp.pretrained_base)
+        base_network = get_base_network(self._cfg.center_net.base_network, pretrained=self._cfg.train.pretrained_base)
         net = get_center_net(self._cfg.center_net.base_network,
                              self._cfg.dataset,
                              base_network=base_network,
@@ -108,8 +107,8 @@ class CenterNetEstimator(BaseEstimator):
                              scale=self._cfg.center_net.scale,
                              topk=self._cfg.center_net.topk,
                              norm_layer=gluon.nn.BatchNorm)
-        if self._cfg.train_hp.resume.strip():
-            net.load_parameters(self._cfg.train_hp.resume.strip())
+        if self._cfg.train.resume.strip():
+            net.load_parameters(self._cfg.train.resume.strip())
         elif os.path.isfile(os.path.join(self._logdir, 'latest.params')):
             net.load_parameters(os.path.join(self._logdir, 'latest.params'))
         else:
@@ -118,7 +117,7 @@ class CenterNetEstimator(BaseEstimator):
                 net.initialize()
 
         # dataloader
-        batch_size = self._cfg.train_hp.batch_size
+        batch_size = self._cfg.train.batch_size
         width, height = self._cfg.center_net.data_shape
         num_class = len(train_dataset.classes)
         batchify_fn = Tuple([Stack() for _ in range(6)])  # stack image, cls_targets, box_targets
@@ -126,12 +125,12 @@ class CenterNetEstimator(BaseEstimator):
             train_dataset.transform(CenterNetDefaultTrainTransform(
                 width, height, num_class=num_class, scale_factor=net.scale)),
             batch_size, True, batchify_fn=batchify_fn, last_batch='rollover',
-            num_workers=self._cfg.train_hp.num_workers)
+            num_workers=self._cfg.train.num_workers)
         val_batchify_fn = Tuple(Stack(), Pad(pad_val=-1))
         val_loader = gluon.data.DataLoader(
             val_dataset.transform(CenterNetDefaultValTransform(width, height)),
-            self._cfg.valid_hp.batch_size, False, batchify_fn=val_batchify_fn, last_batch='keep',
-            num_workers=self._cfg.valid_hp.num_workers)
+            self._cfg.validation.batch_size, False, batchify_fn=val_batchify_fn, last_batch='keep',
+            num_workers=self._cfg.validation.num_workers)
 
         self._train_data = train_loader
         self._val_data = val_loader
@@ -140,25 +139,25 @@ class CenterNetEstimator(BaseEstimator):
         # trainer
         self._net = net
         self._net.collect_params().reset_ctx(ctx)
-        lr_decay = float(self._cfg.train_hp.lr_decay)
-        lr_steps = sorted(self._cfg.train_hp.lr_decay_epoch)
-        lr_decay_epoch = [e - self._cfg.train_hp.warmup_epochs for e in lr_steps]
-        num_batches = len(train_dataset) // self._cfg.train_hp.batch_size
+        lr_decay = float(self._cfg.train.lr_decay)
+        lr_steps = sorted(self._cfg.train.lr_decay_epoch)
+        lr_decay_epoch = [e - self._cfg.train.warmup_epochs for e in lr_steps]
+        num_batches = len(train_dataset) // self._cfg.train.batch_size
         lr_scheduler = LRSequential([
-            LRScheduler('linear', base_lr=0, target_lr=self._cfg.train_hp.lr,
-                        nepochs=self._cfg.train_hp.warmup_epochs, iters_per_epoch=num_batches),
-            LRScheduler(self._cfg.train_hp.lr_mode, base_lr=self._cfg.train_hp.lr,
-                        nepochs=self._cfg.train_hp.epochs - self._cfg.train_hp.warmup_epochs,
+            LRScheduler('linear', base_lr=0, target_lr=self._cfg.train.lr,
+                        nepochs=self._cfg.train.warmup_epochs, iters_per_epoch=num_batches),
+            LRScheduler(self._cfg.train.lr_mode, base_lr=self._cfg.train.lr,
+                        nepochs=self._cfg.train.epochs - self._cfg.train.warmup_epochs,
                         iters_per_epoch=num_batches,
                         step_epoch=lr_decay_epoch,
-                        step_factor=self._cfg.train_hp.lr_decay, power=2),
+                        step_factor=self._cfg.train.lr_decay, power=2),
         ])
 
         for k, v in self._net.collect_params('.*bias').items():
             v.wd_mult = 0.0
         self._trainer = gluon.Trainer(
                     self._net.collect_params(), 'adam',
-                    {'learning_rate': self._cfg.train_hp.lr, 'wd': self._cfg.train_hp.wd,
+                    {'learning_rate': self._cfg.train.lr, 'wd': self._cfg.train.wd,
                      'lr_scheduler': lr_scheduler})
 
         self._save_prefix = os.path.join(self._logdir, net_name)
@@ -172,7 +171,7 @@ class CenterNetEstimator(BaseEstimator):
         wh_metric = mx.metric.Loss('WHL1')
         center_reg_metric = mx.metric.Loss('CenterRegL1')
 
-        for epoch in range(self._cfg.train_hp.start_epoch, self._cfg.train_hp.epochs):
+        for epoch in range(self._cfg.train.start_epoch, self._cfg.train.epochs):
             wh_metric.reset()
             center_reg_metric.reset()
             heatmap_loss_metric.reset()
@@ -183,7 +182,7 @@ class CenterNetEstimator(BaseEstimator):
             for i, batch in enumerate(self._train_data):
                 split_data = [gluon.utils.split_and_load(batch[ind], ctx_list=self._ctx, batch_axis=0) for ind in range(6)]
                 data, heatmap_targets, wh_targets, wh_masks, center_reg_targets, center_reg_masks = split_data
-                batch_size = self._cfg.train_hp.batch_size
+                batch_size = self._cfg.train.batch_size
                 with autograd.record():
                     sum_losses = []
                     heatmap_losses = []
@@ -206,7 +205,7 @@ class CenterNetEstimator(BaseEstimator):
                 heatmap_loss_metric.update(0, heatmap_losses)
                 wh_metric.update(0, wh_losses)
                 center_reg_metric.update(0, center_reg_losses)
-                if self._cfg.train_hp.log_interval and not (i + 1) % self._cfg.train_hp.log_interval:
+                if self._cfg.train.log_interval and not (i + 1) % self._cfg.train.log_interval:
                     name2, loss2 = wh_metric.get()
                     name3, loss3 = center_reg_metric.get()
                     name4, loss4 = heatmap_loss_metric.get()
@@ -219,9 +218,9 @@ class CenterNetEstimator(BaseEstimator):
             name4, loss4 = heatmap_loss_metric.get()
             self._log.info('[Epoch {}] Training cost: {:.3f}, {}={:.3f}, {}={:.3f}, {}={:.3f}'.format(
                 epoch, (time.time()-tic), name2, loss2, name3, loss3, name4, loss4))
-            if (epoch % self._cfg.valid_hp.interval == 0) or \
-               (self._cfg.train_hp.save_interval and epoch % self._cfg.train_hp.save_interval == 0) or \
-               (epoch == self._cfg.train_hp.epochs - 1):
+            if (epoch % self._cfg.validation.interval == 0) or \
+               (self._cfg.train.save_interval and epoch % self._cfg.train.save_interval == 0) or \
+               (epoch == self._cfg.train.epochs - 1):
                 # consider reduce the frequency of validation to save time
                 map_name, mean_ap = self._evaluate()
                 val_msg = '\n'.join(['{}={}'.format(k, v) for k, v in zip(map_name, mean_ap)])
@@ -229,12 +228,12 @@ class CenterNetEstimator(BaseEstimator):
                 current_map = float(mean_ap[-1])
             else:
                 current_map = 0.
-            save_params(current_map, epoch, self._cfg.train_hp.save_interval, self._save_prefix)
+            save_params(current_map, epoch, self._cfg.train.save_interval, self._save_prefix)
 
     def _evaluate(self):
         """Test on validation dataset."""
         self._eval_metric.reset()
-        self._net.flip_test = self._cfg.valid_hp.flip_test
+        self._net.flip_test = self._cfg.validation.flip_test
         mx.nd.waitall()
         self._net.hybridize()
         for batch in self._val_data:
